refactor: :bug: Correct use of context and color change
Fix the use of the provider and change of the dark modal color
