+++ src/main.jsx
@@ -1,15 +1,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
-import { ThemeProvider } from './components/ThemeContext'
 
 // estilos globales
 import './styles/index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-        <ThemeProvider>
-            <App />
-        </ThemeProvider>
+    <App />
   </React.StrictMode>
 )
